Limit dataflow dependencies
