+++ integration/java/src/main/java/com/google/cloud/dataflow/integration/nexmark/PubsubHelper.java
@@ -16,62 +16,61 @@
 
 package com.google.cloud.dataflow.integration.nexmark;
 
-import com.google.api.client.googleapis.json.GoogleJsonResponseException;
-import com.google.api.client.util.BackOff;
-import com.google.api.client.util.BackOffUtils;
-import com.google.api.client.util.Sleeper;
-import com.google.api.services.pubsub.Pubsub;
-import com.google.api.services.pubsub.model.Subscription;
-import com.google.api.services.pubsub.model.Topic;
-import com.google.cloud.dataflow.sdk.util.AttemptBoundedExponentialBackOff;
-import com.google.cloud.dataflow.sdk.util.Transport;
+import com.google.cloud.dataflow.sdk.io.PubsubGrpcClient;
+import com.google.cloud.dataflow.sdk.options.GcpOptions;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 /**
- * Helper for working with pubsub and gcs.
+ * Helper for working with pubsub.
  */
 public class PubsubHelper {
-  /** Underlying pub/sub client. */
-  private final Pubsub pubsubClient;
+  /**
+   * Underlying pub/sub client.
+   */
+  private final PubsubGrpcClient pubsubClient;
 
-  /** Project id. */
+  /**
+   * Project id.
+   */
   private final String project;
 
-  /** Topics we should delete on close. */
+  /**
+   * Topics we should delete on close.
+   */
   private final List<String> createdTopics;
 
-  /** Subscriptions we should delete on close. */
+  /**
+   * Subscriptions we should delete on close.
+   */
   private final List<String> createdSubscriptions;
 
-  /** How to sleep in retry loops. */
-  private final Sleeper sleeper;
-
-  /** How to backoff in retry loops. */
-  private final BackOff backOff;
-
-  private PubsubHelper(Pubsub pubsubClient, String project) {
+  private PubsubHelper(PubsubGrpcClient pubsubClient, String project) {
     this.pubsubClient = pubsubClient;
     this.project = project;
     createdTopics = new ArrayList<>();
     createdSubscriptions = new ArrayList<>();
-    sleeper = Sleeper.DEFAULT;
-    backOff = new AttemptBoundedExponentialBackOff(3, 500);
   }
 
-  /** Create a helper. */
-  public static PubsubHelper create(NexmarkDriver.Options options) throws IOException {
-    return new PubsubHelper(Transport.newPubsubClient(options).build(),
-                            options.getProject());
+  /**
+   * Create a helper.
+   */
+  public static PubsubHelper create(GcpOptions options, String project) throws IOException {
+    return new PubsubHelper(PubsubGrpcClient.newClient(null, null, options), project);
   }
 
-  /** Return full topic name corresponding to short topic name. */
+  /**
+   * Return full topic name corresponding to short topic name.
+   */
   private String fullTopic(String shortTopic) {
     return String.format("projects/%s/topics/%s", project, shortTopic);
   }
 
-  /** Return full subscription name corresponding to short subscription name. */
+  /**
+   * Return full subscription name corresponding to short subscription name.
+   */
   private String fullSubscription(String shortSubscription) {
     return String.format("projects/%s/subscriptions/%s", project, shortSubscription);
   }
@@ -79,48 +78,36 @@ public class PubsubHelper {
   /**
    * Create a topic from short name. Delete it if it already exists. Ensure the topic will be
    * deleted on cleanup. Return full topic name.
+   *
    * @throws InterruptedException
    */
   public String createTopic(String shortTopic) throws IOException, InterruptedException {
     String topic = fullTopic(shortTopic);
-    while (true) {
-      try {
-        NexmarkUtils.console(null, "create topic %s", topic);
-        pubsubClient.projects().topics().create(topic, new Topic()).execute();
-        createdTopics.add(topic);
-        return topic;
-      } catch (GoogleJsonResponseException ex) {
-        NexmarkUtils.console(null, "attempting to cleanup topic %s", topic);
-        pubsubClient.projects().topics().delete(topic).execute();
-        if (!BackOffUtils.next(sleeper, backOff)) {
-          NexmarkUtils.console(null, "too many retries for creating topic %s", topic);
-          throw ex;
-        }
-      }
+    if (topicExists(topic)) {
+      NexmarkUtils.console(null, "attempting to cleanup topic %s", topic);
+      pubsubClient.deleteTopic(topic);
     }
+    NexmarkUtils.console(null, "create topic %s", topic);
+    pubsubClient.createTopic(topic);
+    createdTopics.add(topic);
+    return topic;
   }
 
-  /** Create a topic from short name if it does not already exist. The topic will not be
+  /**
+   * Create a topic from short name if it does not already exist. The topic will not be
    * deleted on cleanup. Return full topic name.
-   * @throws InterruptedException */
+   *
+   * @throws InterruptedException
+   */
   public String createOrReuseTopic(String shortTopic) throws IOException, InterruptedException {
     String topic = fullTopic(shortTopic);
-    while (true) {
-      try {
-        NexmarkUtils.console(null, "create topic %s", topic);
-        pubsubClient.projects().topics().create(topic, new Topic()).execute();
-        return topic;
-      } catch (GoogleJsonResponseException ex) {
-        if (topicExists(shortTopic)) {
-          NexmarkUtils.console(null, "topic %s already exists", topic);
-          return topic;
-        }
-        if (!BackOffUtils.next(sleeper, backOff)) {
-          NexmarkUtils.console(null, "too many retries for creating/reusing topic %s", topic);
-          throw ex;
-        }
-      }
+    if (topicExists(topic)) {
+      NexmarkUtils.console(null, "topic %s already exists", topic);
+      return topic;
     }
+    NexmarkUtils.console(null, "create topic %s", topic);
+    pubsubClient.createTopic(topic);
+    return topic;
   }
 
   /**
@@ -136,48 +123,33 @@ public class PubsubHelper {
     throw new RuntimeException("topic '" + topic + "' does not already exist");
   }
 
-  /** Does topic corresponding to short name exist? */
+  /**
+   * Does topic corresponding to short name exist?
+   */
   public boolean topicExists(String shortTopic) throws IOException {
     String topic = fullTopic(shortTopic);
-    List<Topic> existingTopics =
-        pubsubClient.projects().topics().list("projects/" + project).execute().getTopics();
-    if (existingTopics != null) {
-      for (Topic existingTopic : existingTopics) {
-        if (existingTopic.getName().equals(topic)) {
-          return true;
-        }
-      }
-    }
-    return false;
+    Collection<String> existingTopics = pubsubClient.listTopics("projects/" + project);
+    return existingTopics.contains(topic);
   }
 
   /**
-   * Create subscription from short name. Ensure the subscription will be deleted
-   * on cleanup. Return full subscription name.
+   * Create subscription from short name. Delete subscription if it already exists. Ensure the
+   * subscription will be deleted on cleanup. Return full subscription name.
+   *
    * @throws InterruptedException
    */
   public String createSubscription(String shortTopic, String shortSubscription)
       throws IOException, InterruptedException {
     String topic = fullTopic(shortTopic);
     String subscription = fullSubscription(shortSubscription);
-    while (true) {
-      try {
-        NexmarkUtils.console(null, "create subscription %s", subscription);
-        pubsubClient.projects()
-            .subscriptions()
-            .create(subscription, new Subscription().setTopic(topic).setAckDeadlineSeconds(60))
-            .execute();
-        createdSubscriptions.add(subscription);
-        return subscription;
-      } catch (GoogleJsonResponseException ex) {
-        NexmarkUtils.console(null, "attempting to cleanup subscription %s", subscription);
-        pubsubClient.projects().subscriptions().delete(subscription).execute();
-        if (!BackOffUtils.next(sleeper, backOff)) {
-          NexmarkUtils.console(null, "too many retries for creating subscription %s", subscription);
-          throw ex;
-        }
-      }
+    if (subscriptionExists(shortTopic, shortSubscription)) {
+      NexmarkUtils.console(null, "attempting to cleanup subscription %s", subscription);
+      pubsubClient.deleteSubscription(subscription);
     }
+    NexmarkUtils.console(null, "create subscription %s", subscription);
+    pubsubClient.createSubscription(subscription, topic, 60);
+    createdSubscriptions.add(subscription);
+    return subscription;
   }
 
   /**
@@ -193,34 +165,26 @@ public class PubsubHelper {
     throw new RuntimeException("subscription'" + subscription + "' does not already exist");
   }
 
-  /** Does subscription corresponding to short name exist? */
+  /**
+   * Does subscription corresponding to short name exist?
+   */
   public boolean subscriptionExists(String shortTopic, String shortSubscription)
       throws IOException {
     String topic = fullTopic(shortTopic);
     String subscription = fullSubscription(shortSubscription);
-    List<Subscription> existingSubscriptions =
-        pubsubClient.projects()
-            .subscriptions()
-            .list("projects/" + project)
-            .execute()
-            .getSubscriptions();
-    if (existingSubscriptions != null) {
-      for (Subscription existingSubscription : existingSubscriptions) {
-        if (existingSubscription.getTopic().equals(topic)
-            && existingSubscription.getName().equals(subscription)) {
-          return true;
-        }
-      }
-    }
-    return false;
+    Collection<String> existingSubscriptions =
+        pubsubClient.listSubscriptions("projects/" + project, topic);
+    return existingSubscriptions.contains(subscription);
   }
 
-  /** Delete all the subscriptions and topics we created. */
+  /**
+   * Delete all the subscriptions and topics we created.
+   */
   public void cleanup() {
     for (String subscription : createdSubscriptions) {
       try {
         NexmarkUtils.console(null, "delete subscription %s", subscription);
-        pubsubClient.projects().subscriptions().delete(subscription).execute();
+        pubsubClient.deleteSubscription(subscription);
       } catch (IOException ex) {
         NexmarkUtils.console(null, "could not delete subscription %s", subscription);
       }
@@ -228,7 +192,7 @@ public class PubsubHelper {
     for (String topic : createdTopics) {
       try {
         NexmarkUtils.console(null, "delete topic %s", topic);
-        pubsubClient.projects().topics().delete(topic).execute();
+        pubsubClient.deleteTopic(topic);
       } catch (IOException ex) {
         NexmarkUtils.console(null, "could not delete topic %s", topic);
       }
